20.02.24

+++ TelegramBot/InlineKeyboardHandler.cs
@@ -12,36 +12,42 @@ namespace TelegramBot
         //получаем button_***_data
         public static string InlineKeyboardDataGetting(CallbackQuery callbackQuery)
         {
-            var buttonCallbackData = callbackQuery.Data;
-            return buttonCallbackData;
+            if (update.Message != null && IsGroupChat(update.Message.Chat.Type))
+            {
+                var buttonCallbackData = callbackQuery.Data;
+                return buttonCallbackData;
+            }
         }
 
         //нужно сравнить button_***_data из callbackquery (оставив ***) с текстом в файле
         public static async Task InlineKeyboardActionAsync(CallbackQuery callbackQuery, ITelegramBotClient botClient, long chatId)
         {
-            var filePath = Path.Combine(Directory.GetParent(Environment.CurrentDirectory).Parent.Parent.FullName, "shoppingListData.txt");
-            var items = await File.ReadAllLinesAsync(filePath); 
-            
-            var button = InlineKeyboardDataGetting(callbackQuery); 
-            var clearButtonData = Regex.Replace(button, "_buttonData", "");
-
-            bool found = false;
-
-            for (int i = 0; i < items.Length; i++)
+            if (update.Message != null && IsGroupChat(update.Message.Chat.Type))
             {
-                if (items[i] == clearButtonData)
+                var filePath = Path.Combine(Directory.GetParent(Environment.CurrentDirectory).Parent.Parent.FullName, "shoppingListData.txt");
+                var items = await File.ReadAllLinesAsync(filePath); 
+            
+                var button = InlineKeyboardDataGetting(callbackQuery); 
+                var clearButtonData = Regex.Replace(button, "_buttonData", "");
+
+                bool found = false;
+
+                for (int i = 0; i < items.Length; i++)
                 {
-                    // зачеркиваем при совпадении
-                    items[i] = $"<s>{items[i]}</s>"; 
-                    var updatedFileContent = string.Join(Environment.NewLine, items);
-                    await File.WriteAllTextAsync(filePath, updatedFileContent);
-                    //удаляем сообщение и обновляем список в чате
-                    await botClient.DeleteMessageAsync(chatId, callbackQuery.Message.MessageId);
+                    if (items[i] == clearButtonData)
+                    {
+                        // зачеркиваем при совпадении
+                        items[i] = $"<s>{items[i]}</s>"; 
+                        var updatedFileContent = string.Join(Environment.NewLine, items);
+                        await File.WriteAllTextAsync(filePath, updatedFileContent);
+                        //удаляем сообщение и обновляем список в чате
+                        await botClient.DeleteMessageAsync(chatId, callbackQuery.Message.MessageId);
                     
-                    TelegramBotProgram bot = new TelegramBotProgram();
-                    await bot.ShowShoppingListAsync(botClient, callbackQuery.Message, CancellationToken.None);
+                        TelegramBotProgram bot = new TelegramBotProgram();
+                        await bot.ShowShoppingListAsync(botClient, callbackQuery.Message, CancellationToken.None);
                     
-                    break;
+                        break;
+                    }
                 }
             }
         }
